Made a header container for logo
I updates the gunk.css file and my own pages to accommodate the logo alongside the welcome user part and logout button. I kinda broke it a little as the buttons of the sidenav bar had to have its own section in the css, but it'll be fix it later.